Biomass task (#135)
* feat: migrate hls pipeline to use stac

* exp: continue creating locust replica data

* exp: add nasa api rate limiting

* fix: fix precommit failure

* fix: fix failing tests

* fix: process result of process_tile correctly

* fix: remove redundant rate limiting

* feat: update data scripts and modules (#126)

* feat: update data modules and scripts

* exp: create sen1flood_replica data again

* exp: create sen1flood_replica data again

* exp: create sen1flood_replica data again

* exp: test remove date limiting on open_stac_items

* exp: create locust replica data v2

* exp: create locust replica data v2

* exp: create locust replica data v2

* feat: add masking_strategy to mask_segmentation_map

* exp: create locust replica v2

* exp: test locust replica v2 val split creation

* exp: create locust_new_true_absence again

* exp: create locust_new_pseudo_absence

* exp: create 2022 us cdl data instageo version

* exp: create locust_new_pseudo_absence

* feat: add create_instageo_us_cdl_records module

* exp: create instageo version of us cdl data

* exp: create locust_new_pseudo_absence

* exp: create instageo version of us cdl data

* exp: train on s2 sen1floods replica

* fix: fix bug related NO_DATA_VALUES

* exp: continue creating instageo version of us cdl data

* exp: create instageo version of us cdl data; test cookie fix

* exp: continue creating instageo version of us cdl data; batch_size=10

* exp: continue creating instageo version of us cdl data; use all resources

* exp: continue creating instageo version of us cdl data; use all resources

* exp: continue creating instageo version of us cdl data; use all resources

* exp: continue creating instageo version of us cdl data; use all resources

* exp: continue creating instageo version of us cdl data; use all resources

* exp: continue creating instageo version of us cdl data; use all resources

* exp: continue creating instageo version of us cdl data; use all resources

* exp: continue creating instageo version of us cdl data; batch_size=5

* exp: test AWS_REQUEST_PAYER=requester

* chore: rename BasePointsPipeline as BasePointsDataPipeline

* chore: update sen1floods_replica.sh

* chore: remove unused gdal options

* feat: add raster pipeline

* fix: remove pytz

* fix: fix pystac.Client bug

* fix: improve reprojection code and add test

* tests: add test for hls utils functions

* tests: add more tests for hls utils

* fix: remove unused library

* feat: migrate hls pipeline to use stac

* exp: add nasa api rate limiting

* fix: fix failing tests

* fix: remove redundant rate limiting

* feat: update data modules and scripts

* exp: test remove date limiting on open_stac_items

* exp: create locust replica data v2

* exp: create locust replica v2

* exp: test locust replica v2 val split creation

* exp: create locust_new_true_absence again

* feat: add create_instageo_us_cdl_records module

* exp: create instageo version of us cdl data

* exp: test biomass task with one raster file

* exp: biomass task with ratelimit of 100 and one raster file

* fix: correct precommit issues

* exp: small biomass dataset with 10 raster file

* exp: small biomass dataset with 5 raster file

* exp: small biomass dataset with 3 raster files and ratelimit of 30

* exp: updated biomass dataset with 3 raster files

* fix: fix conflicts

* exp: launch biomass dataset

* fix: correct precommit issue

* exp: launch biomass dataset

* exp: launch biomass dataset

* exp: launch biomass dataset

* exp: launch biomass dataset

* exp: launch biomass dataset with more RAM

* exp: launch biomass dataset with 1% water in chips and 20 as api ratelimit

* exp: launch biomass dataset with 0% water in labels and 20 ratelimit

* exp: launch biomass dataset with 100m resolution and 1% water in labels

* exp: create subset of biomass task with 30 rasters

* exp: create subset of biomass task with 50 rasters

* exp: create subset of biomass task with 100 rasters

* exp: create subset of biomass task with 150 rasters

* fix: Rebased biomass onto hls-stac-migration to resolve conflicts

* exp: full biomass dataset with 30m resolution - updated

* exp: full biomass dataset with 30m resolution - updated

* fix: rebased with hls-stac-migration

* exp: full biomass dataset with 30m resolution - updated

* exp: biomass dataset with 100 rasters - updated

* exp: biomass dataset with 100 rasters - updated

* exp: biomass subset with 5 rasters and 10 cloud cover

* exp: biomass subset with 100 rasters and 10 cloud cover

* exp: biomass subset with 3 rasters and 10 cloud cover

* exp: biomass subset with 2 rasters and 10 cloud cover

* exp: biomass subset with 2 rasters and 10 cloud cover

* exp: biomass subset with 1 raster and 10 cloud cover

* exp: launch entire biomass dataset with 100m res

* exp: biomass dataset with 5 rasters and 30m res

* exp: biomass dataset with 7 rasters

* exp: biomass dataset with 10 rasters

* exp: biomass 10 rasters for 100m resolution

* exp: biomass 30 rasters for 100m resolution

* exp: entire biomass dataset 30m resolution

* exp: launch 2022 dataset

* exp: prepare biomass dataset for 2022

* exp: prepare biomass dataset for 2022

* exp: preprocess biomass 2022 dataset

* exp : 2022 biomass dataset with 5 rasters

* exp: test paralel dataset creation

* exp: parallel dataset creation part2

* exp: parallel biomass dataset part 3

* exp: parallel biomass dataset part 4

* exp: parallel biomass dataset part 5

* exp: parallel biomass dataset part 6

* feat: add script to gather parallel biomass dataset

* feat: improve script to gather parallel biomass dataset

* exp: gather parallel biomass dataset

* fix: correct pre-commit issues

* fix: correct pre-commit

* fix : improving Readability

* docs: update readme file

* Rebased onto develop

* docs: update README file

* docs: update biomass part in README file

* exp: parallel dataset for biomass 2022 part1

* exp: parallel dataset for biomass 2022 part2

* exp: parallel dataset for biomass 2022 part2

* exp: parallel dataset for biomass 2022 part2

---------

+++ instageo/data/data_pipeline.py
@@ -32,6 +32,7 @@ import geopandas as gpd
 import mgrs
 import numpy as np
 import pandas as pd
+import rasterio  # noqa: F401
 import xarray as xr
 from pyproj import Transformer
 from pystac_client import Client
@@ -76,9 +77,9 @@ def mask_segmentation_map(
         The segmentation map after masking
     """
     if masking_strategy == "each":
-        valid_mask = (chip != chip_no_data_value).any(dim="band").astype(np.uint8)
+        valid_mask = (chip != chip_no_data_value).any(dim="band").astype(np.uint16)
     elif masking_strategy == "any":
-        valid_mask = (chip != chip_no_data_value).all(dim="band").astype(np.uint8)
+        valid_mask = (chip != chip_no_data_value).all(dim="band").astype(np.uint16)
     else:
         raise ValueError(f"Invalid masking strategy: {masking_strategy}")
     seg_no_data_value = NO_DATA_VALUES.SEG_MAP
